Moving LayerVariable to its own file
