+++ src/engine/training.ts
@@ -20,9 +20,10 @@ import {NotImplementedError, RuntimeError, ValueError} from '../errors';
 import * as losses from '../losses';
 import * as Metrics from '../metrics';
 import * as optimizers from '../optimizers';
-import {LayerVariable, LossOrMetricFn, Shape} from '../types';
+import {LossOrMetricFn, Shape} from '../types';
 import {count, singletonOrArray, unique} from '../utils/generic_utils';
 import {range} from '../utils/math_utils';
+import {LayerVariable} from '../variables';
 
 import {execute, FeedDict} from './executor';
 import {Container, ContainerConfig} from './topology';
