+++ src/layers/convolutional.ts
@@ -24,9 +24,10 @@ import {NotImplementedError, ValueError} from '../errors';
 import {getInitializer, Initializer, InitializerIdentifier, serializeInitializer} from '../initializers';
 import {getRegularizer, Regularizer, RegularizerIdentifier, serializeRegularizer} from '../regularizers';
 import {Kwargs, Shape} from '../types';
-import {DType, LayerVariable} from '../types';
+import {DType} from '../types';
 import {convOutputLength, deconvLength, normalizeArray} from '../utils/conv_utils';
 import * as generic_utils from '../utils/generic_utils';
+import {LayerVariable} from '../variables';
 // tslint:enable:max-line-length
 
 /**
