+++ src/layers/normalization.ts
@@ -22,9 +22,9 @@ import {ValueError} from '../errors';
 import {getInitializer, Initializer, InitializerIdentifier, serializeInitializer} from '../initializers';
 import {getRegularizer, Regularizer, RegularizerIdentifier, serializeRegularizer} from '../regularizers';
 import {Kwargs, Shape} from '../types';
-import {LayerVariable} from '../types';
 import * as generic_utils from '../utils/generic_utils';
 import {arrayProd, range} from '../utils/math_utils';
+import {LayerVariable} from '../variables';
 
 // tslint:enable:max-line-length
